requirements.txt
google-generativeai>=0.3.0
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0
requests>=2.28.0
beautifulsoup4>=4.11.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0